tests: add storage manual mixin
This lets tests have finer control over local housekeeping.
